Remove all references to ath_info
It's a dangerous program that should be distributed separately from
MadWifi, with its own README full of serious warnings.

ath_info is not even using MadWifi, it writes to the hardware registers
directly.  ath_info is definitely not needed to use MadWifi.  It could
help with the development, but so could Wireshark and many other
programs that are not part of MadWifi.